+++ anvil_consortium_manager/tests/test_views.py
@@ -1269,6 +1269,107 @@ class AccountDetailTest(TestCase):
         self.assertIn("show_reactivate_button", context)
         self.assertTrue(context["show_reactivate_button"])
 
+    def test_context_unlinked_users_no_unlinked_user(self):
+        account = factories.AccountFactory.create()
+        self.client.force_login(self.user)
+        response = self.client.get(self.get_url(account.uuid))
+        context = response.context_data
+        self.assertIn("unlinked_users", context)
+        self.assertEqual(len(context["unlinked_users"]), 0)
+
+    def test_context_unlinked_users_one_unlinked_user(self):
+        account = factories.AccountFactory.create()
+        user = User.objects.create_user(username="test_unlinked", password="test")
+        account.unlinked_users.add(user)
+        self.client.force_login(self.user)
+        response = self.client.get(self.get_url(account.uuid))
+        context = response.context_data
+        self.assertIn("unlinked_users", context)
+        self.assertEqual(len(context["unlinked_users"]), 1)
+        self.assertIn(user, context["unlinked_users"])
+        self.assertIn(str(user), response.content.decode())
+
+    def test_context_unlinked_users_two_unlinked_users(self):
+        account = factories.AccountFactory.create()
+        user_1 = User.objects.create_user(username="test_unlinked_1", password="test")
+        user_2 = User.objects.create_user(username="test_unlinked_2", password="test")
+        account.unlinked_users.add(user_1, user_2)
+        self.client.force_login(self.user)
+        response = self.client.get(self.get_url(account.uuid))
+        context = response.context_data
+        self.assertIn("unlinked_users", context)
+        self.assertEqual(len(context["unlinked_users"]), 2)
+        self.assertIn(user_1, context["unlinked_users"])
+        self.assertIn(user_2, context["unlinked_users"])
+        self.assertIn(str(user_1), response.content.decode())
+        self.assertIn(str(user_2), response.content.decode())
+
+    def test_context_show_unlink_button_linked_account(self):
+        """An is_inactive flag is included in the context."""
+        account = factories.AccountFactory.create(verified=True)
+        edit_user = User.objects.create_user(username="edit", password="test")
+        edit_user.user_permissions.add(
+            Permission.objects.get(codename=models.AnVILProjectManagerAccess.STAFF_VIEW_PERMISSION_CODENAME),
+            Permission.objects.get(codename=models.AnVILProjectManagerAccess.STAFF_EDIT_PERMISSION_CODENAME),
+        )
+        self.client.force_login(edit_user)
+        response = self.client.get(self.get_url(account.uuid))
+        context = response.context_data
+        self.assertIn("show_unlink_button", context)
+        self.assertTrue(context["show_unlink_button"])
+        self.assertContains(
+            response, reverse("anvil_consortium_manager:accounts:unlink", kwargs={"uuid": account.uuid})
+        )
+
+    def test_context_show_unlink_button_linked_account_view_permission(self):
+        """An is_inactive flag is included in the context."""
+        account = factories.AccountFactory.create(verified=True)
+        self.client.force_login(self.user)
+        response = self.client.get(self.get_url(account.uuid))
+        context = response.context_data
+        self.assertIn("show_unlink_button", context)
+        self.assertTrue(context["show_unlink_button"])
+        self.assertNotContains(
+            response, reverse("anvil_consortium_manager:accounts:unlink", kwargs={"uuid": account.uuid})
+        )
+
+    def test_context_show_unlink_button_unlinked_account(self):
+        """An is_inactive flag is included in the context."""
+        account = factories.AccountFactory.create(verified=False)
+        edit_user = User.objects.create_user(username="edit", password="test")
+        edit_user.user_permissions.add(
+            Permission.objects.get(codename=models.AnVILProjectManagerAccess.STAFF_VIEW_PERMISSION_CODENAME),
+            Permission.objects.get(codename=models.AnVILProjectManagerAccess.STAFF_EDIT_PERMISSION_CODENAME),
+        )
+        self.client.force_login(edit_user)
+        response = self.client.get(self.get_url(account.uuid))
+        context = response.context_data
+        self.assertIn("show_unlink_button", context)
+        self.assertFalse(context["show_unlink_button"])
+        self.assertNotContains(
+            response, reverse("anvil_consortium_manager:accounts:unlink", kwargs={"uuid": account.uuid})
+        )
+
+    def test_context_show_unlink_button_previously_linked(self):
+        """An is_inactive flag is included in the context."""
+        account = factories.AccountFactory.create(verified=True)
+        account.unlinked_users.add(self.user)
+        account.user = None
+        account.save()
+        edit_user = User.objects.create_user(username="edit", password="test")
+        edit_user.user_permissions.add(
+            Permission.objects.get(codename=models.AnVILProjectManagerAccess.STAFF_VIEW_PERMISSION_CODENAME),
+            Permission.objects.get(codename=models.AnVILProjectManagerAccess.STAFF_EDIT_PERMISSION_CODENAME),
+        )
+        self.client.force_login(edit_user)
+        response = self.client.get(self.get_url(account.uuid))
+        context = response.context_data
+        self.assertIn("show_unlink_button", context)
+        self.assertFalse(context["show_unlink_button"])
+        self.assertNotContains(
+            response, reverse("anvil_consortium_manager:accounts:unlink", kwargs={"uuid": account.uuid})
+        )
+
     def test_template_verified_account(self):
         """The template renders with a verified account."""
         obj = factories.AccountFactory.create(verified=True)
@@ -3921,6 +4022,227 @@ class AccountReactivateTest(AnVILAPIMockTestMixin, TestCase):
         messages = [m.message for m in get_messages(response.wsgi_request)]
         self.assertEqual(len(messages), 1)
         self.assertEqual(views.AccountReactivate.message_already_active, str(messages[0]))
+
+
+class AccountUnlinkUserTest(TestCase):
+    def setUp(self):
+        """Set up test class."""
+        super().setUp()
+        self.factory = RequestFactory()
+        # Create a user with both view and edit permissions.
+        self.user = User.objects.create_user(username="test", password="test")
+        self.user.user_permissions.add(
+            Permission.objects.get(codename=models.AnVILProjectManagerAccess.STAFF_VIEW_PERMISSION_CODENAME)
+        )
+        self.user.user_permissions.add(
+            Permission.objects.get(codename=models.AnVILProjectManagerAccess.STAFF_EDIT_PERMISSION_CODENAME)
+        )
+
+    def get_url(self, *args):
+        """Get the url for the view being tested."""
+        return reverse("anvil_consortium_manager:accounts:unlink", args=args)
+
+    def get_view(self):
+        """Return the view being tested."""
+        return views.AccountUnlinkUser.as_view()
+
+    def test_view_redirect_not_logged_in(self):
+        "View redirects to login view when user is not logged in."
+        # Need a client for redirects.
+        uuid = uuid4()
+        response = self.client.get(self.get_url(uuid))
+        self.assertRedirects(response, resolve_url(settings.LOGIN_URL) + "?next=" + self.get_url(uuid))
+
+    def test_status_code_with_user_permission(self):
+        """Returns successful response code."""
+        instance = factories.AccountFactory.create(verified=True)
+        self.client.force_login(self.user)
+        response = self.client.get(self.get_url(instance.uuid))
+        self.assertEqual(response.status_code, 200)
+
+    def test_access_with_view_permission(self):
+        """Raises permission denied if user has only view permission."""
+        user_with_view_perm = User.objects.create_user(username="test-other", password="test-other")
+        user_with_view_perm.user_permissions.add(
+            Permission.objects.get(codename=models.AnVILProjectManagerAccess.STAFF_VIEW_PERMISSION_CODENAME)
+        )
+        uuid = uuid4()
+        request = self.factory.get(self.get_url(uuid))
+        request.user = user_with_view_perm
+        with self.assertRaises(PermissionDenied):
+            self.get_view()(request, uuid=uuid)
+
+    def test_access_with_limited_view_permission(self):
+        """Raises permission denied if user has limited view permission."""
+        uuid = uuid4()
+        user = User.objects.create_user(username="test-limited", password="test-limited")
+        user.user_permissions.add(
+            Permission.objects.get(codename=models.AnVILProjectManagerAccess.VIEW_PERMISSION_CODENAME)
+        )
+        request = self.factory.get(self.get_url(uuid))
+        request.user = user
+        with self.assertRaises(PermissionDenied):
+            self.get_view()(request, uuid=uuid)
+
+    def test_access_without_user_permission(self):
+        """Raises permission denied if user has no permissions."""
+        user_no_perms = User.objects.create_user(username="test-none", password="test-none")
+        uuid = uuid4()
+        request = self.factory.get(self.get_url(uuid))
+        request.user = user_no_perms
+        with self.assertRaises(PermissionDenied):
+            self.get_view()(request, uuid=uuid)
+
+    def test_get_context_data(self):
+        """Context data is correct."""
+        instance = factories.AccountFactory.create(verified=True)
+        self.client.force_login(self.user)
+        response = self.client.get(self.get_url(instance.uuid))
+        self.assertIn("object", response.context_data)
+        self.assertEqual(instance, response.context_data["object"])
+
+    def test_view_with_invalid_object(self):
+        """Returns a 404 when the object doesn't exist."""
+        uuid = uuid4()
+        request = self.factory.get(self.get_url(uuid))
+        request.user = self.user
+        with self.assertRaises(Http404):
+            self.get_view()(request, uuid=uuid)
+
+    def test_unlinks_user(self):
+        """Successful post request unlinks the user from the account."""
+        instance = factories.AccountFactory.create(verified=True)
+        self.client.force_login(self.user)
+        response = self.client.post(self.get_url(instance.uuid), {"submit": ""})
+        self.assertEqual(response.status_code, 302)
+        instance.refresh_from_db()
+        self.assertEqual(instance.user, None)
+        self.assertEqual(instance.verified_email_entry, None)
+
+    def test_unlinks_user_not_verified(self):
+        """Successful post request unlinks the user from the account."""
+        instance = factories.AccountFactory.create()
+        user = factories.UserFactory.create()
+        instance.user = user
+        instance.save()
+        self.client.force_login(self.user)
+        response = self.client.post(self.get_url(instance.uuid), {"submit": ""})
+        self.assertEqual(response.status_code, 302)
+        instance.refresh_from_db()
+        self.assertEqual(instance.user, None)
+        self.assertEqual(instance.verified_email_entry, None)
+
+    def test_adds_user_to_unlinked_users(self):
+        """A record is added to unlinked_users."""
+        instance = factories.AccountFactory.create(verified=True)
+        verified_email_entry = instance.verified_email_entry
+        user = instance.user
+        self.client.force_login(self.user)
+        response = self.client.post(self.get_url(instance.uuid), {"submit": ""})
+        self.assertEqual(response.status_code, 302)
+        instance.refresh_from_db()
+        # User link was archived.
+        self.assertIn(user, instance.unlinked_users.all())
+        self.assertEqual(models.AccountUserArchive.objects.count(), 1)
+        archive = models.AccountUserArchive.objects.first()
+        self.assertEqual(archive.account, instance)
+        self.assertEqual(archive.user, user)
+        self.assertIsNotNone(archive.created)
+        self.assertEqual(archive.verified_email_entry, verified_email_entry)
+
+    def test_adds_user_to_unlinked_users_not_verified(self):
+        """A record is added to unlinked_users."""
+        instance = factories.AccountFactory.create()
+        user = factories.UserFactory.create()
+        instance.user = user
+        instance.save()
+        self.client.force_login(self.user)
+        response = self.client.post(self.get_url(instance.uuid), {"submit": ""})
+        self.assertEqual(response.status_code, 302)
+        instance.refresh_from_db()
+        # User link was archived.
+        self.assertIn(user, instance.unlinked_users.all())
+        self.assertEqual(models.AccountUserArchive.objects.count(), 1)
+        archive = models.AccountUserArchive.objects.first()
+        self.assertEqual(archive.account, instance)
+        self.assertEqual(archive.user, user)
+        self.assertIsNotNone(archive.created)
+        self.assertIsNone(archive.verified_email_entry)
+
+    def test_can_add_second_user_to_unlinked_users(self):
+        """A record is added to unlinked_users."""
+        old_user = User.objects.create_user(username="old_user", password="test")
+        instance = factories.AccountFactory.create(verified=True)
+        instance.unlinked_users.add(old_user)
+        user = instance.user
+        self.client.force_login(self.user)
+        response = self.client.post(self.get_url(instance.uuid), {"submit": ""})
+        self.assertEqual(response.status_code, 302)
+        instance.refresh_from_db()
+        # User link was archived.
+        self.assertIn(user, instance.unlinked_users.all())
+        self.assertEqual(models.AccountUserArchive.objects.count(), 2)
+        models.AccountUserArchive.objects.get(account=instance, user=old_user)
+        models.AccountUserArchive.objects.get(account=instance, user=user)
+
+    def test_get_account_has_no_user_redirect(self):
+        instance = factories.AccountFactory.create()
+        self.client.force_login(self.user)
+        response = self.client.get(self.get_url(instance.uuid), {"submit": ""})
+        self.assertRedirects(response, instance.get_absolute_url())
+
+    def test_get_account_no_user_message(self):
+        # A message is included.
+        instance = factories.AccountFactory.create()
+        self.client.force_login(self.user)
+        response = self.client.get(self.get_url(instance.uuid), {"submit": ""}, follow=True)
+        self.assertEqual(response.status_code, 200)
+        messages = [m.message for m in get_messages(response.wsgi_request)]
+        self.assertEqual(len(messages), 1)
+        self.assertEqual(views.AccountUnlinkUser.message_no_user, str(messages[0]))
+        instance.refresh_from_db()
+        self.assertIsNone(instance.user)
+
+    def test_post_account_has_no_user_redirect(self):
+        instance = factories.AccountFactory.create()
+        self.client.force_login(self.user)
+        response = self.client.post(self.get_url(instance.uuid), {"submit": ""})
+        self.assertRedirects(response, instance.get_absolute_url())
+
+    def test_post_account_no_user_message(self):
+        # A message is included.
+        instance = factories.AccountFactory.create()
+        self.client.force_login(self.user)
+        response = self.client.post(self.get_url(instance.uuid), {"submit": ""}, follow=True)
+        self.assertEqual(response.status_code, 200)
+        messages = [m.message for m in get_messages(response.wsgi_request)]
+        self.assertEqual(len(messages), 1)
+        self.assertEqual(views.AccountUnlinkUser.message_no_user, str(messages[0]))
+        instance.refresh_from_db()
+        self.assertIsNone(instance.user)
+
+    def test_success_message(self):
+        instance = factories.AccountFactory.create(verified=True)
+        self.client.force_login(self.user)
+        response = self.client.post(self.get_url(instance.uuid), {"submit": ""}, follow=True)
+        self.assertEqual(response.status_code, 200)
+        messages = [m.message for m in get_messages(response.wsgi_request)]
+        self.assertEqual(len(messages), 1)
+        self.assertEqual(views.AccountUnlinkUser.success_message, str(messages[0]))
+
+    def test_only_unlinks_specified_instance(self):
+        """View only deletes the specified pk."""
+        instance = factories.AccountFactory.create(verified=True)
+        other_instance = factories.AccountFactory.create(verified=True)
+        other_user = other_instance.user
+        self.client.force_login(self.user)
+        response = self.client.post(self.get_url(instance.uuid), {"submit": ""})
+        self.assertEqual(response.status_code, 302)
+        self.assertEqual(models.Account.objects.count(), 2)
+        instance.refresh_from_db()
+        other_instance.refresh_from_db()
+        self.assertEqual(instance.user, None)
+        self.assertEqual(other_instance.user, other_user)
 
 
 class AccountAutocompleteTest(TestCase):